Update README.md
When implementing this, our team mistook 'LOGSTASH_TYPE' being set to 'ansible' to be a kind of data input or formatting rather than a destination index in Elk, which is what it is. I propose updating the description to include this information, as I know of several engineers who have not realised this and run into difficulties because of it.